Simplify evaluate code.
